+++ packages/rdf/lib/append.js
@@ -3,6 +3,16 @@ import { localFetch } from './localFetch/localFetch.js'
 import { applyOptions } from './metadata/applyOptions.js'
 
 class MetadataAppend extends Transform {
+  /**
+   * @param {import('barnard59-core').Context} context
+   * @param {string | undefined} basePath
+   * @param {string} input
+   * @param {{
+   *   dateCreated?: import('./metadata/namedDateLiterals.js').NamedDateLiteral,
+   *   dateModified?: import('./metadata/namedDateLiterals.js').NamedDateLiteral,
+   *   graph?: string | import('@rdfjs/types').NamedNode
+   * }} options
+   */
   constructor(context, basePath, input, options) {
     super({ objectMode: true })
     this.context = context
@@ -11,17 +21,25 @@ class MetadataAppend extends Transform {
     this.options = options
   }
 
+  /**
+   * @param {import('@rdfjs/types').Quad} chunk
+   * @param {string} encoding
+   * @param {import('stream').TransformCallback} callback
+   */
   _transform(chunk, encoding, callback) {
     callback(null, chunk)
   }
 
+  /**
+   * @param {import('stream').TransformCallback} callback
+   */
   async _flush(callback) {
     try {
       const { quadStream, metadata } = await localFetch.call(this.context, this.input, this.basePath)
       for (const quad of await applyOptions(quadStream, metadata, this.options)) {
         this.push(quad)
       }
-    } catch (err) {
+    } catch (/** @type {any} */ err) {
       this.destroy(err)
     } finally {
       callback()
@@ -29,6 +47,16 @@ class MetadataAppend extends Transform {
   }
 }
 
+/**
+ * @this import('barnard59-core').Context
+ * @param {object} [options]
+ * @param {string | undefined} [options.input]
+ * @param {string} [options.basepath]
+ * @param {import('./metadata/namedDateLiterals.js').NamedDateLiteral} [options.dateModified]
+ * @param {import('./metadata/namedDateLiterals.js').NamedDateLiteral} [options.dateCreated]
+ * @param {*} [options.graph]
+ * @return {Promise<Transform>}
+ */
 async function append({
   input,
   basepath,