+++ packages/rdf/test/voidStats.test.js
@@ -1,22 +1,19 @@
-import { equal, strictEqual } from 'assert'
+import { strictEqual } from 'node:assert'
+import { expect } from 'chai'
 import assertThrows from 'assert-throws-async'
 import getStream from 'get-stream'
 import { isDuplexStream as isDuplex } from 'is-stream'
 import rdf from 'barnard59-env'
 import { Readable } from 'readable-stream'
-import toCanonical from 'rdf-dataset-ext/toCanonical.js'
-import * as ns from '../lib/namespaces.js'
-import voidStats from '../lib/voidStats.js'
+import voidStatsUnbound from '../lib/voidStats.js'
+
+const voidStats = voidStatsUnbound.bind({ env: rdf })
 
 const ex = rdf.namespace('http://example.org/')
 /**
  * https://www.w3.org/TR/void/#statistics
  */
 describe('metadata.voidStats', () => {
-  it('should be a factory', () => {
-    strictEqual(typeof voidStats, 'function')
-  })
-
   it('throws an error if no argument is given', async () => {
     await assertThrows(async () => {
       await voidStats()
@@ -33,15 +30,15 @@ describe('metadata.voidStats', () => {
 
   it('includes counts at the end of the stream', async () => {
     const data = [
-      rdf.quad(ex.bob, ns.rdf.type, ex.Person),
-      rdf.quad(ex.alice, ns.rdf.type, ex.Person),
+      rdf.quad(ex.bob, rdf.ns.rdf.type, ex.Person),
+      rdf.quad(ex.alice, rdf.ns.rdf.type, ex.Person),
       rdf.quad(ex.bob, ex.knows, ex.alice),
       rdf.quad(ex.alice, ex.name, rdf.literal('Alice')),
     ]
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.triples, rdf.literal('4', ns.xsd.integer)),
-      rdf.quad(ex.dataset, ns._void.entities, rdf.literal('2', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.triples, rdf.literal('4', rdf.ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns._void.entities, rdf.literal('2', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -51,18 +48,15 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result.slice(4)),
-      toCanonical(expectedMetadata),
-    )
+    expect(result.slice(4)).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('returns zero counts for no data', async () => {
     const data = []
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.triples, rdf.literal('0', ns.xsd.integer)),
-      rdf.quad(ex.dataset, ns._void.entities, rdf.literal('0', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.triples, rdf.literal('0', rdf.ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -72,10 +66,7 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result),
-      toCanonical(expectedMetadata),
-    )
+    expect(result).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('returns zero counts for 0 classes', async () => {
@@ -84,9 +75,9 @@ describe('metadata.voidStats', () => {
       rdf.quad(ex.alice, ex.name, rdf.literal('Alice')),
     ]
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.triples, rdf.literal('2', ns.xsd.integer)),
-      rdf.quad(ex.dataset, ns._void.entities, rdf.literal('0', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.triples, rdf.literal('2', rdf.ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -96,18 +87,15 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result.slice(2)),
-      toCanonical(expectedMetadata),
-    )
+    expect(result.slice(2)).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('uses the named-graph given as parameter', async () => {
     const data = []
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset, ex.metadata),
-      rdf.quad(ex.dataset, ns._void.triples, rdf.literal('0', ns.xsd.integer), ex.metadata),
-      rdf.quad(ex.dataset, ns._void.entities, rdf.literal('0', ns.xsd.integer), ex.metadata),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset, ex.metadata),
+      rdf.quad(ex.dataset, rdf.ns._void.triples, rdf.literal('0', rdf.ns.xsd.integer), ex.metadata),
+      rdf.quad(ex.dataset, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer), ex.metadata),
     ]
     const inputStream = Readable.from(data)
 
@@ -118,10 +106,7 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result),
-      toCanonical(expectedMetadata),
-    )
+    expect(result).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('does not include total counts with includeTotals: false', async () => {
@@ -130,7 +115,7 @@ describe('metadata.voidStats', () => {
       rdf.quad(ex.alice, ex.name, rdf.literal('Alice')),
     ]
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
     ]
     const inputStream = Readable.from(data)
 
@@ -141,29 +126,26 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result.slice(2)),
-      toCanonical(expectedMetadata),
-    )
+    expect(result.slice(2)).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('describes counts for class partitions', async () => {
     const data = [
-      rdf.quad(ex.a_1, ns.rdf.type, ex.A),
-      rdf.quad(ex.a_2, ns.rdf.type, ex.A),
-      rdf.quad(ex.b_1, ns.rdf.type, ex.B),
-      rdf.quad(ex.c_1, ns.rdf.type, ex.C),
+      rdf.quad(ex.a_1, rdf.ns.rdf.type, ex.A),
+      rdf.quad(ex.a_2, rdf.ns.rdf.type, ex.A),
+      rdf.quad(ex.b_1, rdf.ns.rdf.type, ex.B),
+      rdf.quad(ex.c_1, rdf.ns.rdf.type, ex.C),
     ]
     const partition1 = rdf.namedNode('http://example.org/dataset/classPartition/0')
     const partition2 = rdf.namedNode('http://example.org/dataset/classPartition/1')
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.classPartition, partition1),
-      rdf.quad(ex.dataset, ns._void.classPartition, partition2),
-      rdf.quad(partition1, ns._void.class, ex.A),
-      rdf.quad(partition1, ns._void.entities, rdf.literal('2', ns.xsd.integer)),
-      rdf.quad(partition2, ns._void.class, ex.C),
-      rdf.quad(partition2, ns._void.entities, rdf.literal('1', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.classPartition, partition1),
+      rdf.quad(ex.dataset, rdf.ns._void.classPartition, partition2),
+      rdf.quad(partition1, rdf.ns._void.class, ex.A),
+      rdf.quad(partition1, rdf.ns._void.entities, rdf.literal('2', rdf.ns.xsd.integer)),
+      rdf.quad(partition2, rdf.ns._void.class, ex.C),
+      rdf.quad(partition2, rdf.ns._void.entities, rdf.literal('1', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -175,10 +157,7 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result.slice(4)),
-      toCanonical(expectedMetadata),
-    )
+    expect(result.slice(4)).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('describe counts for class partitions with no matches', async () => {
@@ -186,13 +165,13 @@ describe('metadata.voidStats', () => {
     const partition1 = rdf.namedNode('http://example.org/dataset/classPartition/0')
     const partition2 = rdf.namedNode('http://example.org/dataset/classPartition/1')
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.classPartition, partition1),
-      rdf.quad(ex.dataset, ns._void.classPartition, partition2),
-      rdf.quad(partition1, ns._void.class, ex.A),
-      rdf.quad(partition1, ns._void.entities, rdf.literal('0', ns.xsd.integer)),
-      rdf.quad(partition2, ns._void.class, ex.C),
-      rdf.quad(partition2, ns._void.entities, rdf.literal('0', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.classPartition, partition1),
+      rdf.quad(ex.dataset, rdf.ns._void.classPartition, partition2),
+      rdf.quad(partition1, rdf.ns._void.class, ex.A),
+      rdf.quad(partition1, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer)),
+      rdf.quad(partition2, rdf.ns._void.class, ex.C),
+      rdf.quad(partition2, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -204,10 +183,7 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result),
-      toCanonical(expectedMetadata),
-    )
+    expect(result).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('describe counts for property partitions', async () => {
@@ -220,13 +196,13 @@ describe('metadata.voidStats', () => {
     const partition1 = rdf.namedNode('http://example.org/dataset/propertyPartition/0')
     const partition2 = rdf.namedNode('http://example.org/dataset/propertyPartition/1')
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.propertyPartition, partition1),
-      rdf.quad(ex.dataset, ns._void.propertyPartition, partition2),
-      rdf.quad(partition1, ns._void.property, ex.p_1),
-      rdf.quad(partition1, ns._void.entities, rdf.literal('2', ns.xsd.integer)),
-      rdf.quad(partition2, ns._void.property, ex.p_3),
-      rdf.quad(partition2, ns._void.entities, rdf.literal('1', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.propertyPartition, partition1),
+      rdf.quad(ex.dataset, rdf.ns._void.propertyPartition, partition2),
+      rdf.quad(partition1, rdf.ns._void.property, ex.p_1),
+      rdf.quad(partition1, rdf.ns._void.entities, rdf.literal('2', rdf.ns.xsd.integer)),
+      rdf.quad(partition2, rdf.ns._void.property, ex.p_3),
+      rdf.quad(partition2, rdf.ns._void.entities, rdf.literal('1', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -238,10 +214,7 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result.slice(4)),
-      toCanonical(expectedMetadata),
-    )
+    expect(result.slice(4)).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('describe counts for property partitions with no matches', async () => {
@@ -249,13 +222,13 @@ describe('metadata.voidStats', () => {
     const partition1 = rdf.namedNode('http://example.org/dataset/propertyPartition/0')
     const partition2 = rdf.namedNode('http://example.org/dataset/propertyPartition/1')
     const expectedMetadata = [
-      rdf.quad(ex.dataset, ns.rdf.type, ns._void.Dataset),
-      rdf.quad(ex.dataset, ns._void.propertyPartition, partition1),
-      rdf.quad(ex.dataset, ns._void.propertyPartition, partition2),
-      rdf.quad(partition1, ns._void.property, ex.p_1),
-      rdf.quad(partition1, ns._void.entities, rdf.literal('0', ns.xsd.integer)),
-      rdf.quad(partition2, ns._void.property, ex.p_3),
-      rdf.quad(partition2, ns._void.entities, rdf.literal('0', ns.xsd.integer)),
+      rdf.quad(ex.dataset, rdf.ns.rdf.type, rdf.ns._void.Dataset),
+      rdf.quad(ex.dataset, rdf.ns._void.propertyPartition, partition1),
+      rdf.quad(ex.dataset, rdf.ns._void.propertyPartition, partition2),
+      rdf.quad(partition1, rdf.ns._void.property, ex.p_1),
+      rdf.quad(partition1, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer)),
+      rdf.quad(partition2, rdf.ns._void.property, ex.p_3),
+      rdf.quad(partition2, rdf.ns._void.entities, rdf.literal('0', rdf.ns.xsd.integer)),
     ]
     const inputStream = Readable.from(data)
 
@@ -267,18 +240,15 @@ describe('metadata.voidStats', () => {
 
     const result = await getStream.array(inputStream.pipe(sut))
 
-    equal(
-      toCanonical(result),
-      toCanonical(expectedMetadata),
-    )
+    expect(result).to.deep.contain.all.members(expectedMetadata)
   })
 
   it('accepts string parameters', async () => {
     const data = [
-      rdf.quad(ex.a_1, ns.rdf.type, ex.A),
-      rdf.quad(ex.a_2, ns.rdf.type, ex.A),
-      rdf.quad(ex.b_1, ns.rdf.type, ex.B),
-      rdf.quad(ex.c_1, ns.rdf.type, ex.C),
+      rdf.quad(ex.a_1, rdf.ns.rdf.type, ex.A),
+      rdf.quad(ex.a_2, rdf.ns.rdf.type, ex.A),
+      rdf.quad(ex.b_1, rdf.ns.rdf.type, ex.B),
+      rdf.quad(ex.c_1, rdf.ns.rdf.type, ex.C),
       rdf.quad(ex.a, ex.p_1, ex.b),
       rdf.quad(ex.a, ex.p_1, ex.b),
       rdf.quad(ex.a, ex.p_2, ex.b),